Brigandine iron/steel fix ## About The Pull Request
There were two lightweight brigandine recipes, one with iron and one with steel, that made the same item. Lightweight Brig was intended to be a steel upgrade to studded leather. This brings the recipes in line with that vision.

Removed LB iron recipe, LB now smelts into steel as intended.

## Testing Evidence

<!-- It's mandatory to test your PR. Provide images, clips or description of how you tested your changes where possible. -->

## Why It's Good For The Game

Patches an exploit and more clearly defines light armor progression.

<!-- By contributing to this codebase, you confirm that any code and sprites you provide are legal to share and will be licensed under the terms specified in README.md — AGPLv3 for code and CC-BY-SA 3.0 for assets, unless otherwise stated. You acknowledge that the project maintainers are under no obligation to remove any materials that do not violate these licenses. -->